backend/requirements.txt
Flask
Flask-CORS
python-dotenv # .env 파일 로드용
# google-generativeai (Gemini API 사용 시 필요) 